old-tests/submission/Microsoft/selection/selectionStartEnd.htm

<!DOCTYPE HTML>
<html>
    <head>
        <title id="desc">HTML5 Selection: Set selectionStart and selectionEnd on a text field</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="common.js"></script>
        <script type="text/javascript">
            function RunTest()
            {
                var input1 = document.getElementById("input1");
                var expectedResult = "text";

                input1.selectionStart = 5;
                input1.selectionEnd = 9;

                checkDefaultSelectionAttributes();

                var selectionText = input1.value.substring(input1.selectionStart, input1.selectionEnd);
                assert_equals(selectionText, expectedResult);
            }
        </script>
    </head>
    <body onload="test(RunTest);">
        <input style="WIDTH: 500px" id="input1" value="Some text in an input control" type="text" />
        <p>Select some text in the input element by setting selectionStart and selectionEnd</p>
    </body>
</html>
